Hide file references in quick chat, and fix tests
