+++ src/vs/workbench/contrib/chat/browser/chatEditor.ts
@@ -59,7 +59,7 @@ export class ChatEditor extends EditorPane {
 			scopedInstantiationService.createInstance(
 				ChatWidget,
 				{ resource: true },
-				{},
+				{ supportsFileReferences: true },
 				{
 					listForeground: editorForeground,
 					listBackground: editorBackground,
