+++ src/vs/workbench/contrib/chat/browser/chatViewPane.ts
@@ -91,7 +91,7 @@ export class ChatViewPane extends ViewPane implements IChatViewPane {
 			this._widget = this._register(scopedInstantiationService.createInstance(
 				ChatWidget,
 				{ viewId: this.id },
-				{},
+				{ supportsFileReferences: true },
 				{
 					listForeground: SIDE_BAR_FOREGROUND,
 					listBackground: this.getBackgroundColor(),
